remove unused aescache, move pow algorithm into hash.h
