+++ src/main.h
@@ -25,7 +25,6 @@
 #include "tinyformat.h"
 #include "txmempool.h"
 #include "uint256.h"
-#include "aescache.h"
 
 #include <algorithm>
 #include <exception>
@@ -503,8 +502,6 @@ extern CCoinsViewCache *pcoinsTip;
 extern CBlockTreeDB *pblocktree;
 
 extern int minerStopFlag;
-
-extern CBlockAesCache *aesCache;
 
 
 #endif // BITCOIN_MAIN_H
