+++ src/miner.cpp
@@ -602,8 +602,7 @@ void static BitcoinMiner(CWallet *pwallet, uint32_t minerI, uint32_t minerN, int
                 for (int i = 0; i < 20 && minerStopFlag == 0 && fGenerate && !ShutdownRequested(); i++) {
                     pblock->nNonce = nNonce;
                     nNonce += minerN;
-                    //hash = pblock->FindBestPatternHash(collisions, scratchpad, nThreads, &minerStopFlag);
-                    hash = pblock->GetPoWHash();
+                    hash = pblock->GetHash();
 
 										/*
                     //totalHashes = totalHashes + collisions;
