+++ src/primitives/block.h
@@ -71,10 +71,6 @@ public:
     }
 
     uint256 GetHash() const;
-    uint256 GetPoWHash() const;
-    uint256 GetHashNoCache() const;
-    uint256 GetMidHash() const;
-    //uint256 FindBestPatternHash(int& collisions,char *scratchpad,int nThreads,int* minerStopFlag);
 
     int64_t GetBlockTime() const
     {
